Added -i option to set initial bar fill percentage
There is a problem with using `mouse_state` to fill the bar to the right
amount of pixels when you use the `--initial-percent` option because
`relative_mouse_mode` does not allow you to move the cursor with
`warp_mouse_in_window`. In order to implement `--initial-percent` I
needed to recompute the bar fill value (`fill_pixels`). But in the case
that the cursor is not grabbed, we can take the more efficient route and
simply use mouse_state.